+++ pkg/commands/service/domain/list.go
@@ -1,0 +1,122 @@
+package domain
+
+import (
+	"context"
+	"fmt"
+	"io"
+
+	"github.com/fastly/go-fastly/v12/fastly"
+
+	"github.com/fastly/cli/pkg/argparser"
+	fsterr "github.com/fastly/cli/pkg/errors"
+	"github.com/fastly/cli/pkg/global"
+	"github.com/fastly/cli/pkg/text"
+)
+
+// ListCommand calls the Fastly API to list domains.
+type ListCommand struct {
+	argparser.Base
+	argparser.JSONOutput
+
+	Input          fastly.ListDomainsInput
+	serviceName    argparser.OptionalServiceNameID
+	serviceVersion argparser.OptionalServiceVersion
+}
+
+// NewListCommand returns a usable command registered under the parent.
+func NewListCommand(parent argparser.Registerer, g *global.Data) *ListCommand {
+	c := ListCommand{
+		Base: argparser.Base{
+			Globals: g,
+		},
+	}
+	c.CmdClause = parent.Command("list", "List domains on a Fastly service version")
+
+	// Required.
+	c.RegisterFlag(argparser.StringFlagOpts{
+		Name:        argparser.FlagVersionName,
+		Description: argparser.FlagVersionDesc,
+		Dst:         &c.serviceVersion.Value,
+		Required:    true,
+	})
+
+	// Optional.
+	c.RegisterFlagBool(c.JSONFlag()) // --json
+	c.RegisterFlag(argparser.StringFlagOpts{
+		Name:        argparser.FlagServiceIDName,
+		Description: argparser.FlagServiceIDDesc,
+		Dst:         &g.Manifest.Flag.ServiceID,
+		Short:       's',
+	})
+	c.RegisterFlag(argparser.StringFlagOpts{
+		Action:      c.serviceName.Set,
+		Name:        argparser.FlagServiceName,
+		Description: argparser.FlagServiceNameDesc,
+		Dst:         &c.serviceName.Value,
+	})
+	return &c
+}
+
+// Exec invokes the application logic for the command.
+func (c *ListCommand) Exec(_ io.Reader, out io.Writer) error {
+	if c.Globals.Verbose() && c.JSONOutput.Enabled {
+		return fsterr.ErrInvalidVerboseJSONCombo
+	}
+
+	serviceID, serviceVersion, err := argparser.ServiceDetails(argparser.ServiceDetailsOpts{
+		APIClient:          c.Globals.APIClient,
+		Manifest:           *c.Globals.Manifest,
+		Out:                out,
+		ServiceNameFlag:    c.serviceName,
+		ServiceVersionFlag: c.serviceVersion,
+		VerboseMode:        c.Globals.Flags.Verbose,
+	})
+	if err != nil {
+		c.Globals.ErrLog.AddWithContext(err, map[string]any{
+			"Service ID":      serviceID,
+			"Service Version": fsterr.ServiceVersion(serviceVersion),
+		})
+		return err
+	}
+
+	c.Input.ServiceID = serviceID
+	c.Input.ServiceVersion = fastly.ToValue(serviceVersion.Number)
+
+	o, err := c.Globals.APIClient.ListDomains(context.TODO(), &c.Input)
+	if err != nil {
+		c.Globals.ErrLog.AddWithContext(err, map[string]any{
+			"Service ID":      serviceID,
+			"Service Version": fastly.ToValue(serviceVersion.Number),
+		})
+		return err
+	}
+
+	if ok, err := c.WriteJSON(out, o); ok {
+		return err
+	}
+
+	if !c.Globals.Verbose() {
+		tw := text.NewTable(out)
+		tw.AddHeader("SERVICE", "VERSION", "NAME", "COMMENT")
+		for _, domain := range o {
+			tw.AddLine(
+				fastly.ToValue(domain.ServiceID),
+				fastly.ToValue(domain.ServiceVersion),
+				fastly.ToValue(domain.Name),
+				fastly.ToValue(domain.Comment),
+			)
+		}
+		tw.Print()
+		return nil
+	}
+
+	fmt.Fprintf(out, "Version: %d\n", c.Input.ServiceVersion)
+	for i, domain := range o {
+		fmt.Fprintf(out, "\tDomain %d/%d\n", i+1, len(o))
+		fmt.Fprintf(out, "\t\tName: %s\n", fastly.ToValue(domain.Name))
+		fmt.Fprintf(out, "\t\tComment: %v\n", fastly.ToValue(domain.Comment))
+	}
+	fmt.Fprintln(out)
+
+	return nil
+}