+++ pkg/commands/service/domain/update.go
@@ -1,0 +1,119 @@
+package domain
+
+import (
+	"context"
+	"fmt"
+	"io"
+
+	"github.com/fastly/go-fastly/v12/fastly"
+
+	"4d63.com/optional"
+
+	"github.com/fastly/cli/pkg/argparser"
+	"github.com/fastly/cli/pkg/errors"
+	"github.com/fastly/cli/pkg/global"
+	"github.com/fastly/cli/pkg/text"
+)
+
+// UpdateCommand calls the Fastly API to update domains.
+type UpdateCommand struct {
+	argparser.Base
+	input          fastly.UpdateDomainInput
+	serviceName    argparser.OptionalServiceNameID
+	serviceVersion argparser.OptionalServiceVersion
+	autoClone      argparser.OptionalAutoClone
+
+	NewName argparser.OptionalString
+	Comment argparser.OptionalString
+}
+
+// NewUpdateCommand returns a usable command registered under the parent.
+func NewUpdateCommand(parent argparser.Registerer, g *global.Data) *UpdateCommand {
+	c := UpdateCommand{
+		Base: argparser.Base{
+			Globals: g,
+		},
+	}
+	c.CmdClause = parent.Command("update", "Update a domain on a Fastly service version")
+
+	// Required.
+	c.CmdClause.Flag("name", "Domain name").Short('n').Required().StringVar(&c.input.Name)
+	c.RegisterFlag(argparser.StringFlagOpts{
+		Name:        argparser.FlagVersionName,
+		Description: argparser.FlagVersionDesc,
+		Dst:         &c.serviceVersion.Value,
+		Required:    true,
+	})
+
+	// Optional.
+	c.RegisterAutoCloneFlag(argparser.AutoCloneFlagOpts{
+		Action: c.autoClone.Set,
+		Dst:    &c.autoClone.Value,
+	})
+	c.CmdClause.Flag("comment", "A descriptive note").Action(c.Comment.Set).StringVar(&c.Comment.Value)
+	c.CmdClause.Flag("new-name", "New domain name").Action(c.NewName.Set).StringVar(&c.NewName.Value)
+	c.RegisterFlag(argparser.StringFlagOpts{
+		Name:        argparser.FlagServiceIDName,
+		Description: argparser.FlagServiceIDDesc,
+		Dst:         &g.Manifest.Flag.ServiceID,
+		Short:       's',
+	})
+	c.RegisterFlag(argparser.StringFlagOpts{
+		Action:      c.serviceName.Set,
+		Name:        argparser.FlagServiceName,
+		Description: argparser.FlagServiceNameDesc,
+		Dst:         &c.serviceName.Value,
+	})
+	return &c
+}
+
+// Exec invokes the application logic for the command.
+func (c *UpdateCommand) Exec(_ io.Reader, out io.Writer) error {
+	serviceID, serviceVersion, err := argparser.ServiceDetails(argparser.ServiceDetailsOpts{
+		Active:             optional.Of(false),
+		Locked:             optional.Of(false),
+		AutoCloneFlag:      c.autoClone,
+		APIClient:          c.Globals.APIClient,
+		Manifest:           *c.Globals.Manifest,
+		Out:                out,
+		ServiceNameFlag:    c.serviceName,
+		ServiceVersionFlag: c.serviceVersion,
+		VerboseMode:        c.Globals.Flags.Verbose,
+	})
+	if err != nil {
+		c.Globals.ErrLog.AddWithContext(err, map[string]any{
+			"Service ID":      serviceID,
+			"Service Version": errors.ServiceVersion(serviceVersion),
+		})
+		return err
+	}
+
+	c.input.ServiceID = serviceID
+	c.input.ServiceVersion = fastly.ToValue(serviceVersion.Number)
+
+	// If neither arguments are provided, error with useful message.
+	if !c.NewName.WasSet && !c.Comment.WasSet {
+		return fmt.Errorf("error parsing arguments: must provide either --new-name or --comment to update domain")
+	}
+
+	if c.NewName.WasSet {
+		c.input.NewName = &c.NewName.Value
+	}
+	if c.Comment.WasSet {
+		c.input.Comment = &c.Comment.Value
+	}
+
+	d, err := c.Globals.APIClient.UpdateDomain(context.TODO(), &c.input)
+	if err != nil {
+		c.Globals.ErrLog.AddWithContext(err, map[string]any{
+			"Service ID":      serviceID,
+			"Service Version": fastly.ToValue(serviceVersion.Number),
+			"New Name":        c.NewName.Value,
+			"Comment":         c.Comment.Value,
+		})
+		return err
+	}
+
+	text.Success(out, "Updated domain %s (service %s version %d)", fastly.ToValue(d.Name), fastly.ToValue(d.ServiceID), fastly.ToValue(d.ServiceVersion))
+	return nil
+}